Remove a lot of useless function

+++ srcs/utils/str_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   str_utils.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ambelkac <ambelkac@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jcluzet <jcluzet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/26 18:44:56 by ambelkac          #+#    #+#             */
-/*   Updated: 2021/11/10 18:03:48 by ambelkac         ###   ########.fr       */
+/*   Updated: 2021/11/11 00:30:35 by jcluzet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -59,7 +59,7 @@ int	is_in_set(char c, char const *set)
 	return (0);
 }
 
-size_t	len(const char *str)
+int	len(const char *str)
 {
 	int				i;
